requirements.txt
django==4.1.5
pillow==9.4.0
django-ordered-model==3.6
django-autoslug==1.9.8
django-environ==0.9.0
psycopg2==2.9.5
black==23.1.0
django-debug-toolbar==3.8.1
factory_boy==3.2.1
django-solo==2.0.0
django-cleanup==7.0.0
django-colorfield==0.8.0
djangorestframework==3.14.0
